Split test helpers by layer so model tests carry no client imports
Addresses the Codex review: test_models.py imported a payload factory from the mixed
helpers.py, which eagerly pulled in httpx, AsyncGitHubClient, and the rate-limiting stack,
contradicting the "no transport or client" boundary the model suite is supposed to hold.

Split the shared helpers by layer:
- payloads.py: pure dict payload factories, stdlib-only imports. This is all test_models.py
  needs.
- helpers.py: client-layer material (json_response, make_client, recording_transport,
  governed_client, patch_signed_download, make_zip, first_page, EndpointCase, ENDPOINT_CALLS),
  importing response bodies from payloads, never the reverse.

Repoint every test's payload-factory imports at payloads.py. Document the rule in the package
AGENTS.md layer-ownership section: helper modules obey the same layer boundaries as tests, and
a model test needing something from helpers.py means the helper is misplaced or the test is not
a model test.
